refactor file and post results into their own components
