+++ app/screens/home/index.tsx
@@ -20,9 +20,8 @@ import Account from './account';
 import ChannelList from './channel_list';
 import RecentMentions from './recent_mentions';
 import SavedMessages from './saved_messages';
+import Search from './search';
 import TabBar from './tab_bar';
-
-// import Search from './search';
 
 import type {LaunchProps} from '@typings/launch';
 
@@ -125,11 +124,11 @@ export default function HomeScreen(props: HomeProps) {
                 >
                     {() => <ChannelList {...props}/>}
                 </Tab.Screen>
-                {/* <Tab.Screen
+                <Tab.Screen
                     name={Screens.SEARCH}
                     component={Search}
                     options={{unmountOnBlur: false, lazy: true, tabBarTestID: 'tab_bar.search.tab'}}
-                /> */}
+                />
                 <Tab.Screen
                     name={Screens.MENTIONS}
                     component={RecentMentions}
